requirements.txt
Flask==3.1.0
gunicorn==23.0.0
python-docx==1.1.2
PyMuPDF==1.25.5
